:sparkles: support for alternative input modes (control key)

+++ project/game_engine/manager/displaymanager.h
@@ -14,7 +14,7 @@ class DisplayManager
 {
     GLFWwindow *window;
 
-    float deltaTime = 0.0f;
+    double deltaTime = 0.0f;
     int useWireframe = 0;
     int displayGrayscale = 0;
